Wrapped the Start Terminal into its own component
